backend/requirements.txt
fastapi
uvicorn
python-socketio
requests
openai-whisper
python-dotenv
groq
pyaudio
numpy~=1.26.4
librosa
